More template tweaks
Added codemirror/* and ext CDN references to the individual entry templates for the editor.
These libraries are only used for the entry.tmpl template and aren't loaded otherwise. Also tried to optimize JS library loading slightly using `defer` to delay the disqus.js and analytics.js libs to give codemirror a bit more of a priority.